no empty files, improved summary and test data

+++ bio_assembly_refinement/contig_cleanup.py
@@ -37,6 +37,7 @@ class ContigCleanup:
 				 percent_match=95, 
 				 skip = None,
 				 summary_file="contig_cleanup_summary.txt",
+				 summary_prefix="[contig cleanup]",
 				 debug=False):
 				 
 		''' Constructor '''
@@ -44,8 +45,8 @@ class ContigCleanup:
 		self.working_directory = working_directory if working_directory else os.getcwd()			
 		self.cutoff_contig_length = cutoff_contig_length
 		self.percent_match = percent_match
-		self.alignments = utils.run_nucmer(self.fasta_file, self.fasta_file, self._build_nucmer_filename(), min_percent_id=percent_match)
 		self.summary_file = summary_file
+		self.summary_prefix = summary_prefix
 		self.debug = debug		
 		self.contigs = {}
 		tasks.file_to_dict(self.fasta_file, self.contigs) #Read contig ids and sequences into dict
@@ -64,10 +65,21 @@ class ContigCleanup:
 	
 	def _write_summary(self, small_contigs, contained_contigs):
 		'''Write summary'''
-		text = '~~contig filtration~~\n' + \
-			   'small contigs removed: ' + ",".join(small_contigs) + "\n" \
-			   'contained contigs removed: ' + ",".join(contained_contigs) + "\n"
-		utils.write_text_to_file(text, self.summary_file)
+		header = self.summary_prefix + " cleaning " + self.fasta_file + "\n"
+		utils.write_text_to_file(header, self.summary_file)
+		#small contigs
+		if small_contigs:
+			line = self.summary_prefix + " small contigs removed: " + " ".join(small_contigs) + "\n"
+		else:
+			line = self.summary_prefix + " small contigs removed: " + " (none) \n"
+		utils.write_text_to_file(line, self.summary_file)
+		
+		#contained contigs
+		if contained_contigs:
+			line = self.summary_prefix + " contained contigs removed: " + " ".join(contained_contigs) + "\n"
+		else:
+			line = self.summary_prefix + " contained contigs removed: " + " (none) \n"
+		utils.write_text_to_file(line, self.summary_file)
 				
 		
 	def _build_final_filename(self):
@@ -87,26 +99,27 @@ class ContigCleanup:
 		os.chdir(self.working_directory)
 		small_contigs = set()
 		contained_contigs = set()
-		for id in self.contigs.keys():
-			if not id in self.ids_to_skip:
-				if len(self.contigs[id]) < self.cutoff_contig_length:
-					small_contigs.add(id)
-				else:
-					for algn in self.alignments:
-						if (not algn.is_self_hit()) \
-						   and algn.qry_name == id \
-						   and algn.ref_name != algn.qry_name \
-						   and not algn.ref_name in contained_contigs \
-						   and (algn.hit_length_qry/algn.qry_length) * 100 >= self.percent_match:
-							contained_contigs.add(id)
+		if len(self.contigs) > len(self.ids_to_skip):
+			alignments = utils.run_nucmer(self.fasta_file, self.fasta_file, self._build_nucmer_filename(), min_percent_id=self.percent_match)
+			for id in self.contigs.keys():
+				if not id in self.ids_to_skip:
+					if len(self.contigs[id]) < self.cutoff_contig_length:
+						small_contigs.add(id)
+					else:
+						for algn in alignments:
+							if (not algn.is_self_hit()) \
+							   and algn.qry_name == id \
+							   and algn.ref_name != algn.qry_name \
+							   and not algn.ref_name in contained_contigs \
+							   and (algn.hit_length_qry/algn.qry_length) * 100 >= self.percent_match:
+								contained_contigs.add(id)
 					
-		discard = small_contigs.union(contained_contigs)
-		ids_file = utils.write_ids_to_file(discard, "contig.ids.discard")  
-		tasks.filter(self.fasta_file, self.output_file, ids_file=ids_file, invert=True)		
-		self._write_summary(small_contigs, contained_contigs)
-		
-		if not self.debug:
-			utils.delete(ids_file)
-			utils.delete(self._build_nucmer_filename())
-		
+			discard = small_contigs.union(contained_contigs)
+			ids_file = utils.write_ids_to_file(discard, "contig.ids.discard")  
+			tasks.filter(self.fasta_file, self.output_file, ids_file=ids_file, invert=True)		
+			if not self.debug:
+				utils.delete(ids_file)
+				utils.delete(self._build_nucmer_filename())
+			
+		self._write_summary(small_contigs, contained_contigs)	
 		os.chdir(original_dir)
